VESPER_CUDA/requirements.txt
numpy
mrcfile
numba
pyFFTW
scipy
tqdm
scikit-learn
biopython
BioTEMPy